Move FileDesc to separated header file
